backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
alembic==1.13.2
python-multipart==0.0.9
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.32.3
python-dateutil==2.9.0.post0
apscheduler==3.10.4
jinja2==3.1.4
itsdangerous==2.2.0
pillow==10.4.0
imagehash==4.3.1
ExifRead==3.0.0
python-dotenv==1.1.1

